Add missing type casts in HeaderAssertions
Two header assertions (expires, lastModified) accept int parameter
whereas HttpHeaders class stores those headers as long.
It makes them always fail as Object::equals is used for comparing.

The change casts incoming values to long.

Issues: SPR-17195

+++ spring-test/src/main/java/org/springframework/test/web/reactive/server/HeaderAssertions.java
@@ -182,14 +182,14 @@ public class HeaderAssertions {
 	 * Expect an "Expires" header with the given value.
 	 */
 	public WebTestClient.ResponseSpec expires(int expires) {
-		return assertHeader("Expires", expires, getHeaders().getExpires());
+		return assertHeader("Expires", (long)expires, getHeaders().getExpires());
 	}
 
 	/**
 	 * Expect a "Last-Modified" header with the given value.
 	 */
 	public WebTestClient.ResponseSpec lastModified(int lastModified) {
-		return assertHeader("Last-Modified", lastModified, getHeaders().getLastModified());
+		return assertHeader("Last-Modified", (long)lastModified, getHeaders().getLastModified());
 	}
